Closes make render job take multiple record ids, only download media once #355
